Move ConfigCenter to infra package
