+++ shardingsphere-governance/shardingsphere-governance-core/src/main/java/org/apache/shardingsphere/governance/core/registry/cache/subscriber/ScalingRegistrySubscriber.java
@@ -22,8 +22,8 @@ import org.apache.shardingsphere.governance.core.registry.cache.RegistryCacheMan
 import org.apache.shardingsphere.governance.core.registry.cache.event.StartScalingEvent;
 import org.apache.shardingsphere.governance.core.registry.config.event.rule.RuleConfigurationCachedEvent;
 import org.apache.shardingsphere.governance.core.registry.config.event.rule.SwitchRuleConfigurationEvent;
-import org.apache.shardingsphere.governance.core.config.service.impl.SchemaRulePersistService;
-import org.apache.shardingsphere.governance.core.config.node.SchemaMetadataNode;
+import org.apache.shardingsphere.infra.config.persist.service.impl.SchemaRulePersistService;
+import org.apache.shardingsphere.infra.config.persist.node.SchemaMetadataNode;
 import org.apache.shardingsphere.governance.repository.spi.RegistryCenterRepository;
 import org.apache.shardingsphere.infra.config.RuleConfiguration;
 import org.apache.shardingsphere.infra.eventbus.ShardingSphereEventBus;
