+++ shardingsphere-governance/shardingsphere-governance-core/src/main/java/org/apache/shardingsphere/governance/core/registry/config/watcher/GlobalRuleChangedWatcher.java
@@ -18,7 +18,7 @@
 package org.apache.shardingsphere.governance.core.registry.config.watcher;
 
 import com.google.common.base.Preconditions;
-import org.apache.shardingsphere.governance.core.config.node.GlobalNode;
+import org.apache.shardingsphere.infra.config.persist.node.GlobalNode;
 import org.apache.shardingsphere.governance.core.registry.GovernanceWatcher;
 import org.apache.shardingsphere.governance.core.registry.config.event.rule.GlobalRuleConfigurationsChangedEvent;
 import org.apache.shardingsphere.governance.repository.api.listener.DataChangedEvent;
