+++ shardingsphere-infra/shardingsphere-infra-common/src/main/java/org/apache/shardingsphere/infra/config/persist/service/GlobalPersistService.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.governance.core.config.service;
+package org.apache.shardingsphere.infra.config.persist.service;
 
 /**
  * Global persist service.
