+++ shardingsphere-infra/shardingsphere-infra-common/src/test/java/org/apache/shardingsphere/infra/config/persist/service/impl/DataSourcePersistServiceTest.java
@@ -15,12 +15,11 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.governance.core.config.impl;
+package org.apache.shardingsphere.infra.config.persist.service.impl;
 
 import lombok.SneakyThrows;
-import org.apache.shardingsphere.governance.core.config.service.impl.DataSourcePersistService;
-import org.apache.shardingsphere.governance.repository.spi.RegistryCenterRepository;
 import org.apache.shardingsphere.infra.config.datasource.DataSourceConfiguration;
+import org.apache.shardingsphere.infra.config.persist.repository.ConfigCenterRepository;
 import org.apache.shardingsphere.test.mock.MockedDataSource;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -45,7 +44,7 @@ import static org.mockito.Mockito.when;
 public final class DataSourcePersistServiceTest {
     
     @Mock
-    private RegistryCenterRepository repository;
+    private ConfigCenterRepository repository;
     
     @Test
     public void assertLoad() {
@@ -58,7 +57,7 @@ public final class DataSourcePersistServiceTest {
     
     @SneakyThrows({IOException.class, URISyntaxException.class})
     private String readDataSourceYaml() {
-        return Files.readAllLines(Paths.get(ClassLoader.getSystemResource("yaml/regcenter/data-source.yaml").toURI()))
+        return Files.readAllLines(Paths.get(ClassLoader.getSystemResource("yaml/configcenter/data-source.yaml").toURI()))
                 .stream().filter(each -> !each.startsWith("#")).map(each -> each + System.lineSeparator()).collect(Collectors.joining());
     }
