+++ shardingsphere-infra/shardingsphere-infra-common/src/test/java/org/apache/shardingsphere/infra/config/persist/service/impl/PropertiesPersistServiceTest.java
@@ -15,22 +15,33 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.governance.core.config.node;
+package org.apache.shardingsphere.infra.config.persist.service.impl;
 
+import org.apache.shardingsphere.infra.config.persist.repository.ConfigCenterRepository;
+import org.apache.shardingsphere.infra.config.properties.ConfigurationPropertyKey;
 import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.mockito.Mock;
+import org.mockito.junit.MockitoJUnitRunner;
+
+import java.util.Properties;
 
 import static org.hamcrest.CoreMatchers.is;
 import static org.junit.Assert.assertThat;
+import static org.mockito.Mockito.when;
 
-public final class GlobalNodeTest {
+@RunWith(MockitoJUnitRunner.class)
+public final class PropertiesPersistServiceTest {
+    
+    private static final String PROPS_YAML = ConfigurationPropertyKey.SQL_SHOW.getKey() + ": false\n";
+    
+    @Mock
+    private ConfigCenterRepository repository;
     
     @Test
-    public void assertGetGlobalRuleNodePath() {
-        assertThat(GlobalNode.getGlobalRuleNode(), is("/rules"));
-    }
-    
-    @Test
-    public void assertGetPropsPath() {
-        assertThat(GlobalNode.getPropsPath(), is("/props"));
+    public void assertLoad() {
+        when(repository.get("/props")).thenReturn(PROPS_YAML);
+        Properties actual = new PropertiesPersistService(repository).load();
+        assertThat(actual.get(ConfigurationPropertyKey.SQL_SHOW.getKey()), is(Boolean.FALSE));
     }
 }
