+++ shardingsphere-infra/shardingsphere-infra-common/src/test/resources/yaml/configcenter/data-global-rule.yaml
